refactor(shirelang): refactor output handling and improve code structure #18
Refactor the handling of output in the Shirelang project. The output of the compiler is now stored in a new variable 'shireOutput' instead of 'output'. This change has been propagated throughout the codebase. Additionally, the code structure has been improved by extracting classes 'ShireConsoleView' and 'ShireProcessAdapter' for better readability and maintainability.