refactor(compiler): replace ProcessFuncNode with PostProcessorNode #24
The commit updates the compiler module by removing the `ProcessFuncNode` class and replacing it with the new `PostProcessorNode` class. This change affects the `HobbitHole` and its associated processing functions, as well as the test cases to reflect the use of the new node type. The `PostProcessor` interface has been extended to include the execution of the `PostProcessorNode`.