feat(editor): add smart code insertion feature #24
Added a new feature to insert code at the current cursor position in the editor. This includes a new InsertCodeProcessor and an update to the CodeModifier interface to support smart code insertion. Also, the CodeModifier class has been moved to a new package. The return types of setup, execute, and finish methods in PostProcessor have been changed to Any.