refactor(kotlin): remove unnecessary null-coalescing and adjust class structures #54
This commit simplifies the code by eliminating a null-coalescing operator that is no longer needed in `SonarqubeVariableProvider.kt`. It also restructures the `ShireCoroutineScope` class by moving the `coroutineExceptionHandler` inside the class and adjusting the order of the properties and companion object. Additionally, it updates the issue key in `SonarqubeVariable.kt` for clarity.