refactor(httpclient): change return type and remove scratch file creation #11
The conversion function now returns a `RestClientRequest` directly instead of `Any?`. Additionally, the code responsible for creating and opening a scratch file has been commented out, as it seems to be no longer necessary. A corresponding test case has been added to verify the conversion of a cURL command to a `RestClientRequest`.