docs(http-api-tool): add quick start guide and details on ShireEnv and JsonPath && closed #11
This commit introduces a quick start section to the HTTP API tool documentation and provides further explanations on the `.shireEnv.json` file and how to use JsonPath within Shire. It also adds a new function, `thread`, for multithreaded execution, and updates the pattern functions list with a new `jsonpath` function for JSON data selection.