refactor(runner): extract syntax analysis to separate function #42
This change moves the syntax analysis process into its own function to improve code reusability and readability. The `preAnalysisSyntax` function in `ShireRunCompile.kt` is now responsible for parsing the syntax, which is then utilized in `ShireRunConfigurationProfileState.kt` and `ShireRunner.kt`. This extraction allows for better management of the syntax analysis process and reduces code duplication. Additionally, `ShireRunnerContext` now includes a new parameter for `compiledVariables` to store variables compiled during the process.