feat(testing): add Shire language annotation to test cases and implement shell script runner
This commit adds the @Language("Shire") annotation to several test cases in ShireCompileTest.kt for better language support. It also introduces a new test file, ShireShellRunnerTest.kt, which tests the functionality of filling in environment variables in shell scripts.

In addition, the ShireShellRunner.kt file has been updated to include a method for filling in environment variables in shell scripts, and to execute shell scripts with these filled-in variables.

Some unused imports in CUrlHttpHandler.kt have been removed for code cleanliness.