refactor(marketplace): reorganize code and update button functionality #86
Reorganized the code in the MarketplaceToolWindowFactory class for better readability. Moved the creation of the dataList and columns array to the top of the class. Updated the button functionality to disable cell editing and added a click listener to the ButtonRenderer class. Also updated the install action message in the ShireMainBundle properties file.