refactor(marketplace): update ShireDownloader to use background thread #86
The ShireDownloader class has been refactored to use a background thread for downloading and unzipping files. This change improves the performance and responsiveness of the application. The download and unzip methods have been updated to use a lock to ensure thread safety. Additionally, the ShireDownloader class has been updated to use the DownloadableFileService for downloading files. The downloadFile and unzipFile methods have been removed and replaced with the doDownload method. The ShireMainBundle properties file has also been updated to include a new message for downloading packages.